pyproject.toml
[tool.poetry]
name = "party-up"
version = "0.1.0"
description = ""
authors = ["royce79-creator <connorsboyce07@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
docker = "^5.0.3"
djangorestframework-simplejwt = "^5.1.0"
gunicorn = "^20.1.0"
httpie = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
